refactor : usecase 변경
usecase를 interface와 impl로 분리함.
