+++ domain/src/main/java/com/knichu/domain/usecase/AirPollutionUseCase.kt
@@ -1,41 +1,10 @@
-package com.knichu.domain.usecase
+package com.knichu.domain.useCase
 
-import com.knichu.domain.BuildConfig
-import com.knichu.domain.datastore.WeatherDataStore
-import com.knichu.domain.param.AirPollutionRequestParam
-import com.knichu.domain.repository.AirPollutionRepository
-import com.knichu.domain.repository.CityLocationRepository
 import com.knichu.domain.vo.AirPollutionVO
 import io.reactivex.rxjava3.core.Single
-import javax.inject.Inject
 
-class AirPollutionUseCase @Inject constructor(
-    private val airPollutionRepository: AirPollutionRepository,
-    private val cityLocationRepository: CityLocationRepository,
-) {
-    private val allCityList = cityLocationRepository.getAllCityLocation()
-    private val openWeatherMapApiKey = BuildConfig.OPEN_WEATHER_MAP_API_KEY
+interface AirPollutionUseCase {
+    fun getCurrentPositionAirPollution(lon: Double, lat: Double): Single<AirPollutionVO>
 
-    fun getCurrentPositionAirPollution(lon: Double, lat: Double): Single<AirPollutionVO> {
-        return airPollutionRepository.getAirPollution(
-            AirPollutionRequestParam(
-                lon = lon,
-                lat = lat
-            )
-        )
-    }
-
-    fun getCityPositionAirPollution(city: String): Single<AirPollutionVO> {
-        return allCityList
-            .flatMap { cityLocationVO ->
-                val cityLocationItemVO = cityLocationVO.item?.find { it.cityName == city }
-
-                airPollutionRepository.getAirPollution(
-                    AirPollutionRequestParam(
-                        lon = cityLocationItemVO?.longitude?.toDouble(),
-                        lat = cityLocationItemVO?.latitude?.toDouble()
-                    )
-                )
-            }
-    }
+    fun getCityPositionAirPollution(city: String): Single<AirPollutionVO>
 }
